First example: update dates to be during ACCESS,
use Resource as the example for filtering rather than
Service Provider since it is easier to understand
conceptually, and move the "start" and "end" duration
example first, since it is (and should be) more likely to be
used in practice.
Note that the Plotly and IPython version numbers
changed due to my development/test environment
changing from Docker to Anaconda, but this is
inconsequential.